Add TestProcess WF + GetStrategy
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C22650

+++ CDRAnalysis/Vanrise.Fzero.FraudAnalysis.Entities/Strategy.cs
@@ -8,6 +8,16 @@ namespace Vanrise.Fzero.FraudAnalysis.Entities
 {
     public class Strategy
     {
+        //public string Description { get; set; }
+
+        //public int UserId { get; set; }
+
+        //public DateTime CreationDate { get; set; }
+
+        //public string Name { get; set; }
+
+        //public bool IsDefault { get; set; }
+
         public List<StrategyCriteria> Criterias { get; set; }
 
         public List<StrategyLevel> Levels { get; set; }
@@ -31,7 +41,7 @@ namespace Vanrise.Fzero.FraudAnalysis.Entities
     {
         public int CriteriaId { get; set; }
 
-        public int Percentage { get; set; }
+        public decimal Percentage { get; set; }
     }
 }
